+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/models/MediaStreamingMetadata.java
@@ -29,7 +29,7 @@ public class MediaStreamingMetadata extends MediaStreamingPackageBase {
     /*
      * The length.
      */
-    private final double length;
+    private final int length;
 
     /**
      * The constructor
@@ -40,7 +40,7 @@ public class MediaStreamingMetadata extends MediaStreamingPackageBase {
      * @param channels The channels.
      * @param length The length.
      */
-    public MediaStreamingMetadata(String mediaSubscriptionId, String encoding, int sampleRate, int channels, double length) {
+    public MediaStreamingMetadata(String mediaSubscriptionId, String encoding, int sampleRate, int channels, int length) {
         this.mediaSubscriptionId = mediaSubscriptionId;
         this.encoding = encoding;
         this.sampleRate = sampleRate;
@@ -89,7 +89,7 @@ public class MediaStreamingMetadata extends MediaStreamingPackageBase {
      *
      * @return the length value.
      */
-    public double getLength() {
+    public int getLength() {
         return length;
     }
 }
